pyproject.toml
[project]
name = "linkedin-icebreaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "google-search-results>=2.4.2",
    "gradio>=6.4.0",
    "langchain>=1.2.7",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "typer>=0.21.1",
]
